modify counter action
